add cli project & move source generator to own project
