+++ source-generator/Cli/Program.cs
@@ -1,0 +1,6 @@
+﻿var input = args[0];
+var output = args[1];
+
+var generatedCode = File.ReadAllText(input);
+//parse json from generated code
+File.WriteAllText(output, generatedCode);
