+++ source-generator/CodeGen/ApplicationModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain;
+﻿namespace CodeGen;
 
 public class ApplicationModel
 {
