+++ source-generator/CodeGen/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "Profile 1": {
+      "commandName": "DebugRoslynComponent",
+      "targetProject": "..\\Domain\\Domain.csproj"
+    }
+  }
+}
