+++ source-generator/WebApp/analyzer.config.json
@@ -1,0 +1,3 @@
+{
+  "schema.json": "ControllerTemplate.schema"
+}
