+++ source-generator/source-generator.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApp", "WebApp\WebApp.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "Domain\Domain.csproj", "{E23CBC32-8B95-4873-9983-B595CA6AA04D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cli", "Cli\Cli.csproj", "{7517FB40-609F-473D-8646-747E79A470B7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeGen", "CodeGen\CodeGen.csproj", "{38BB404B-552B-4BB8-8081-A04C1085BB58}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,8 +25,19 @@ Global
 		{E23CBC32-8B95-4873-9983-B595CA6AA04D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E23CBC32-8B95-4873-9983-B595CA6AA04D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E23CBC32-8B95-4873-9983-B595CA6AA04D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7517FB40-609F-473D-8646-747E79A470B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7517FB40-609F-473D-8646-747E79A470B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7517FB40-609F-473D-8646-747E79A470B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7517FB40-609F-473D-8646-747E79A470B7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{38BB404B-552B-4BB8-8081-A04C1085BB58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{38BB404B-552B-4BB8-8081-A04C1085BB58}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{38BB404B-552B-4BB8-8081-A04C1085BB58}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{38BB404B-552B-4BB8-8081-A04C1085BB58}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {03C0BEDA-217B-4EB7-9A03-B0163FB19E72}
 	EndGlobalSection
 EndGlobal
